Fix dialyzer
