+++ lib/boombox/mp4.ex
@@ -55,7 +55,7 @@ defmodule Boombox.MP4 do
   @spec link_output(
           String.t(),
           [Boombox.force_transcoding()],
-          Boombox.Bin.track_builders(),
+          Boombox.InternalBin.track_builders(),
           Membrane.ChildrenSpec.t()
         ) :: Ready.t()
   def link_output(location, opts, track_builders, spec_builder) do
