+++ lib/boombox/webrtc.ex
@@ -110,7 +110,7 @@ defmodule Boombox.WebRTC do
 
   @spec link_output(
           [Boombox.force_transcoding()],
-          Boombox.Bin.track_builders(),
+          Boombox.InternalBin.track_builders(),
           Membrane.ChildrenSpec.t(),
           webrtc_sink_new_tracks(),
           State.t()
@@ -126,7 +126,7 @@ defmodule Boombox.WebRTC do
 
   @spec handle_output_tracks_negotiated(
           [Boombox.force_transcoding()],
-          Boombox.Bin.track_builders(),
+          Boombox.InternalBin.track_builders(),
           Membrane.ChildrenSpec.t(),
           webrtc_sink_new_tracks(),
           State.t()
